refactor(server): extract response sending into server_response.go
Move sendResponse and SendMessage out of server.go into server_response.go.
Same package, mechanical move.
